finished strings exercise

+++ src/shapes/Measurable.java
@@ -1,6 +1,6 @@
 package shapes;
 
 public interface Measurable {
-    int getPerimeter();
-    int getArea();
+    double getPerimeter();
+    double getArea();
 }
